Add redeem code hook
