+++ src/pages/Forever/index.tsx
@@ -5,14 +5,12 @@ import { HedvigLogo } from 'components/icons/HedvigLogo'
 import { useCurrentLocale } from 'components/utils/CurrentLocale'
 import { Page } from 'components/utils/Page'
 import { SessionContainer } from 'containers/SessionContainer'
-import { useRedeemCodeV2Mutation } from 'data/graphql'
-import { FormikHelpers } from 'formik'
 import React from 'react'
 import Helmet from 'react-helmet-async'
-import { RouteComponentProps, useHistory } from 'react-router'
+import { RouteComponentProps } from 'react-router'
 import { useTextKeys } from 'utils/hooks/useTextKeys'
-import { captureSentryError } from 'utils/sentry-client'
-import { RedeemCode, RedeemCodeFormValue } from './components/RedeemCode'
+import { RedeemCode } from './components/RedeemCode'
+import { useRedeemCode } from './useRedeemCode'
 
 type ForeverProps = RouteComponentProps<{
   code: string
@@ -54,77 +52,9 @@ export const Forever: React.FC<ForeverProps> = ({
     params: { code },
   },
 }) => {
-  const history = useHistory()
   const currentLocale = useCurrentLocale()
-  const [redeemCode] = useRedeemCodeV2Mutation()
-  const handleSubmit = async (
-    form: RedeemCodeFormValue,
-    actions: FormikHelpers<RedeemCodeFormValue>,
-  ) => {
-    try {
-      const result = await redeemCode({ variables: { code: form.code } })
-      if (!result) {
-        return
-      }
-
-      if (result.data?.redeemCodeV2.__typename === 'SuccessfulRedeemResult') {
-        // TODO redirect to correct success screens
-        history.push(`/${currentLocale}/new-member`)
-        return
-      }
-      if (result.data?.redeemCodeV2.__typename === 'CannotRedeemOwnCampaign') {
-        actions.setErrors({ code: 'FOREVER_ERROR_CANNOT_REDEEM_OWN_CAMPAIGN' })
-        return
-      }
-      if (
-        result.data?.redeemCodeV2.__typename ===
-        'CampaignCannotBeCombinedWithExisting'
-      ) {
-        actions.setErrors({
-          code: 'FOREVER_ERROR_CAMPAIGN_CANNOT_BE_COMBINED_WITH_EXISTING',
-        })
-        return
-      }
-      if (result.data?.redeemCodeV2.__typename === 'CampaignHasExpired') {
-        actions.setErrors({ code: 'FOREVER_ERROR_CAMPAIGN_HAS_EXPIRED' })
-        return
-      }
-      if (
-        result.data?.redeemCodeV2.__typename ===
-        'MemberIsNotEligibleForCampaign'
-      ) {
-        actions.setErrors({
-          code: 'FOREVER_ERROR_CAMPAIGN_CANNOT_BE_COMBINED_WITH_EXISTING',
-        })
-        return
-      }
-
-      if (result.data?.redeemCodeV2.__typename === 'CampaignDoesNotExist') {
-        actions.setErrors({ code: 'FOREVER_ERROR_CAMPAIGN_DOES_NOT_EXIST' })
-        return
-      }
-
-      if (
-        result.data?.redeemCodeV2.__typename ===
-        'CannotRedeemCampaignFromDifferentMarket'
-      ) {
-        actions.setErrors({
-          code: 'FOREVER_ERROR_CANNOT_REDEEM_CAMPAIGN_FROM_DIFFERENT_MARKET',
-        })
-        return
-      }
-
-      if (result.errors && result.errors.length > 0) {
-        // TODO handle errors
-        actions.setErrors({ code: 'FOREVER_ERROR_GENERIC' })
-        return
-      }
-    } catch (e) {
-      captureSentryError(e)
-    }
-  }
-
   const textKeys = useTextKeys()
+  const { handleSubmit } = useRedeemCode()
   return (
     <>
       <Helmet>
